[network] Move peer_manager related setup to dedicated builder.
Move all setup logic and datastructures associated with PeerManager into a dedicated PeerManagerBuilder.  This reduces the footprint of network_builder and facilitates a structured, phased building process for the network components.
